+++ CHANGELOG.md
@@ -51,6 +51,7 @@ Line wrap the file at 100 chars.                                              Th
 - Make login field keep previous value when submitting an incorrect account number in desktop app.
 - Decrease the time it takes to connect to WireGuard relays by sending an ICMP packet immediately.
 - Pause API interactions when the daemon has not been used for 3 days.
+- Simplified output of `mullvad status` command.
 
 ### Fixed
 - Fix the sometimes incorrect time added text after adding time to the account.